requirements.txt
alembic==1.7.5
aniso8601==9.0.1
boto3==1.20.29
botocore==1.23.29
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
razorpay==1.2.0
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.7
Werkzeug==2.0.2
